test_api_local/requirements.txt
# HTTP client and environment management
requests>=2.31.0
python-dotenv>=1.0.0
urllib3>=2.0.0
pytest>=8.2.0

# Install parent project in editable mode to access core.logger
-e ..
